requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
flake8==3.8.3
idna==2.9
importlib-metadata==1.6.1
mccabe==0.6.1
multidict==4.7.6
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.13.0
websockets==8.1
yarl==1.4.2
zipp==3.1.0
